+++ R/stat_cor.R
@@ -25,8 +25,8 @@ NULL
 #'@param label.x,label.y \code{numeric} Coordinates (in data units) to be used
 #'  for absolute positioning of the label. If too short they will be recycled.
 #'@param output.type character One of "expression", "latex" or "text".
-#'@param digits integer indicating the number of decimal places (round) or
-#'  significant digits (signif) to be used.
+#'@param digits,r.digits,p.digits integer indicating the number of decimal places (round) or
+#'  significant digits (signif) to be used for the correlation coefficient and the p-value, respectively..
 #'@param ... other arguments to pass to \code{\link[ggplot2]{geom_text}} or
 #'  \code{\link[ggplot2]{geom_label}}.
 #'@param na.rm If FALSE (the default), removes missing values with a warning. If
@@ -67,7 +67,7 @@ stat_cor <- function(mapping = NULL, data = NULL,
                      method = "pearson", cor.coef.name = c("R", "rho", "tau"), label.sep = ", ",
                      label.x.npc = "left", label.y.npc = "top",
                      label.x = NULL, label.y = NULL, output.type = "expression",
-                     digits = 2,
+                     digits = 2, r.digits = digits, p.digits = digits,
                      geom = "text", position = "identity",  na.rm = FALSE, show.legend = NA,
                     inherit.aes = TRUE, ...) {
   parse <- ifelse(output.type == "expression", TRUE, FALSE)
@@ -78,7 +78,7 @@ stat_cor <- function(mapping = NULL, data = NULL,
     params = list(label.x.npc  = label.x.npc , label.y.npc  = label.y.npc,
                   label.x = label.x, label.y = label.y, label.sep = label.sep,
                   method = method, output.type = output.type, digits = digits,
-                  cor.coef.name = cor.coef.name,
+                  r.digits = r.digits, p.digits = p.digits, cor.coef.name = cor.coef.name,
                   parse = parse, na.rm = na.rm, ...)
   )
 }
@@ -90,7 +90,7 @@ StatCor<- ggproto("StatCor", Stat,
 
                   compute_group = function(data, scales, method, label.x.npc, label.y.npc,
                                            label.x, label.y, label.sep, output.type, digits,
-                                           cor.coef.name)
+                                           r.digits, p.digits, cor.coef.name)
                     {
                     if (length(unique(data$x)) < 2) {
                       # Not enough data to perform test
@@ -100,7 +100,8 @@ StatCor<- ggproto("StatCor", Stat,
                     .test <- .cor_test(
                       data$x, data$y, method = method, label.sep = label.sep,
                       output.type = output.type, digits = digits,
-                      cor.coef.name
+                      r.digits = r.digits, p.digits = p.digits,
+                      cor.coef.name = cor.coef.name
                       )
                     # Returns a data frame with label: x, y, hjust, vjust
                     .label.pms <- .label_params(data = data, scales = scales,
@@ -118,16 +119,16 @@ StatCor<- ggproto("StatCor", Stat,
 # Correlation test
 #::::::::::::::::::::::::::::::::::::::::
 # Returns a data frame: estimatel|p.value|method|label
-.cor_test <- function(x, y, method = "pearson", label.sep = ", ", output.type = "expression", digits = 2,
-                      cor.coef.name = "R"){
+.cor_test <- function(x, y, method = "pearson", label.sep = ", ", output.type = "expression",
+                      digits = 2, r.digits = digits, p.digits = digits, cor.coef.name = "R"){
 
   .cor <- suppressWarnings(stats::cor.test(x, y, method = method,  use = "complete.obs"))
   estimate <- p.value <- p <- r <- rr <-  NULL
   z <- data.frame(estimate = .cor$estimate, p.value = .cor$p.value, method = method) %>%
     mutate(
-      r = signif(estimate, digits),
-      rr = signif(estimate^2, digits),
-      p = signif(p.value, digits)
+      r = signif(estimate, r.digits),
+      rr = signif(estimate^2, r.digits),
+      p = signif(p.value, p.digits)
     )
 
 
@@ -170,8 +171,8 @@ StatCor<- ggproto("StatCor", Stat,
 
     # Default label
     pvaltxt <- ifelse(pval < 2.2e-16, "p < 2.2e-16",
-                      paste("p =", signif(pval, digits)))
-    cortxt <- paste0("R = ", signif(.cor$estimate, digits),
+                      paste("p =", signif(pval, p.digits)))
+    cortxt <- paste0("R = ", signif(.cor$estimate, r.digits),
                      label.sep,  pvaltxt)
     z$label <- cortxt
   }